feat(grpc_server): add code to setup basic grpc server without dht and exp
Has no expiration or DHT at the moment with this version of the server
