+++ market_server/Cargo.toml
@@ -9,6 +9,8 @@ edition = "2021"
 market_proto = { path = "../market_proto" }
 tonic = { version = "0.11.0" }
 tokio = { version = "1.36.0", features = ["full"] }
+anyhow = { version = "1.0.81" }
+clap = { version = "4.5.3", features = ["derive"] }
 
 [build-dependencies]
 tonic-build = "0.11.0"
